chore: clean up development test page for better developer experience
